aggiornamento collisioni proiettili
uccide i proiettili che collidono anche con le auto, ogni tanto crasha, da fixare.

+++ v24/proiettile.c
@@ -33,10 +33,13 @@ void moveProiettile(int *pipe_fd, PipeData *shooter,int id) {
 }
 
 // uccide il processo proiettile corrispondente all' id passato
-void uccidiProiettile( pid_t *array_pid_proiettili, int id_proiettile){ 
+void uccidiProiettile( pid_t *array_pid_proiettili, int id_proiettile){
+	if((id_proiettile != -1) && (array_pid_proiettili[id_proiettile] != 0))
+	{ 
 		kill(array_pid_proiettili[id_proiettile], SIGKILL);
 		waitpid(array_pid_proiettili[id_proiettile],NULL,0);
 		array_pid_proiettili[id_proiettile]=0;
+	}
 }
