+++ v24/proiettileNemico.c
@@ -34,9 +34,12 @@ void proiettileNemico(int *pipe_fd, PipeData *shooter,int id) {
 
 
 void uccidiProiettileNemico( pid_t *array_pid_proiettili, int id_proiettile){
+	if((id_proiettile != -1) && (array_pid_proiettili[id_proiettile] != 0))
+	{
 		kill(array_pid_proiettili[id_proiettile], SIGKILL);
 		waitpid(array_pid_proiettili[id_proiettile],NULL,0);
 		array_pid_proiettili[id_proiettile]=0;
+	}
 }
